Committing a publicly usable version.
The "MATLAB-library" directory contains a subset of my collection of generally useful functions for MATLAB, some of which are used in this project.

The "example model" contains the SIMION workbench example.iob which loads the model example.pa0, for which the source code in example.lua will be run (a copy of "script to copy and name as each model.lua"). The .lua file contains model-dependent constants that need to be edited for each model, e.g. related to the number of electrodes, resistor chains, size of the simulation volume and the detector location(s).

Signed-off-by: Erik Månsson